rename `registerInteractionCommands.ts` -> `registerApplicationCommands.ts`
