+++ src/reciple/registerApplicationCommands.ts
